requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
PyPDF2==3.0.1
python-docx==0.8.11
